+++ Tests/MapboxMapsTests/Viewport/States/Overview/Random/OverviewViewportStateOptions+Random.swift
@@ -4,7 +4,7 @@ extension OverviewViewportStateOptions {
     static func random() -> Self {
         return OverviewViewportStateOptions(
             geometry: Point(.random()),
-            padding: .random(),
+            coordinatesPadding: .random(),
             bearing: .random(.random(in: 0..<360)),
             pitch: .random(.random(in: 0...80)),
             animationDuration: .random(in: 0..<10))